add pipe telemetry-context
